refactor: do not remove `WORKDIR` `CLEANUP` is set to false
This was something that I did not see it coming soon.

+++ src/util_functions.sh
@@ -73,6 +73,11 @@ function create_ota() {
 }
 
 function cleanup() {
+  if [[ "${CLEANUP}" != 'true' ]]; then
+    echo -e "Cleanup is disabled. Exiting...\n"
+    return
+  fi
+
   echo "Cleaning up..."
   rm -rf "${WORKDIR}"
   unset "${KEYS[@]}"
